+++ firmware/boron-ins-fsm/src/stateMachine.cpp
@@ -24,8 +24,8 @@ unsigned long state0_occupant_detection_timer = STATE0_OCCUPANT_DETECTION_TIMER;
 unsigned long state1_max_time = STATE1_MAX_TIME;
 unsigned long state2_max_duration = STATE2_MAX_DURATION;
 unsigned long state3_max_stillness_time = STATE3_MAX_STILLNESS_TIME;
-unsigned long state3_max_long_stillness_time =
-    STATE3_MAX_STILLNESS_TIME;  // By default, we use the same value for max_stillness_time and max_long_stillness_time
+// By default, we use the same value for max_stillness_time and max_long_stillness_time
+unsigned long state3_max_long_stillness_time = STATE3_MAX_STILLNESS_TIME;
 int resetReason = System.resetReason();
 // record whether an alert has been sent within the same session
 bool hasDurationAlertBeenSent;
@@ -34,6 +34,9 @@ bool hasStillnessAlertBeenSent;
 // using pointers so that the value pointed to be max_stillness_time will be the correct values of state3_max_stillness_time or
 // state3_max_long_stillness_time even if they change due to a console function call during the session
 unsigned long *max_stillness_time = &state3_max_stillness_time;
+// the total number of alerts published while in state 2 or state 3
+// the number of alerts generated in a session should start at 0, resetting to 0 when in state 0
+unsigned long number_of_alerts_published = 0;
 
 std::queue<int> stateQueue;
 std::queue<int> reasonQueue;
@@ -139,6 +142,8 @@ void state0_idle() {
     hasStillnessAlertBeenSent = false;
     // set to compare against the shorter stillness threshold
     max_stillness_time = &state3_max_stillness_time;
+    // set the total number of alerts generated to 0
+    number_of_alerts_published = 0;
 
     // do stuff in the state
     digitalWrite(D2, LOW);
@@ -221,6 +226,8 @@ void state2_duration() {
     checkDoor = checkIM();
     // this returns 0.0 if the INS has no new data to transmit
     checkINS = checkINS3331();
+    // will contain the data for a Stillness Alert; max 622 chars as per Particle docs
+    char alertMessage[622];
 
     // do stuff in the state
     digitalWrite(D3, HIGH);
@@ -248,7 +255,9 @@ void state2_duration() {
         Log.warn("See duration alert, remaining in state2_duration after alert publish");
         saveStateChangeOrAlert(2, 4);
         Log.error("Duration Alert!!");
-        Particle.publish("Duration Alert", "duration alert", PRIVATE);
+        number_of_alerts_published += 1;  // increment the number of alerts published
+        snprintf(alertMessage, sizeof(alertMessage), "{\"numberOfAlertsPublished\": %lu}", number_of_alerts_published);
+        Particle.publish("Duration Alert", alertMessage, PRIVATE);
         hasDurationAlertBeenSent = true;
         state2_duration_timer = millis();
         stateHandler = state2_duration;
@@ -264,6 +273,9 @@ void state3_stillness() {
     // scan inputs
     doorData checkDoor;
     filteredINSData checkINS;
+    // will contain the data for a Stillness Alert; max 622 chars as per Particle docs
+    char alertMessage[622];
+
     // this returns the previous door event value until a new door event is received
     // on code boot up it initializes to returning INITIAL_DOOR_STATUS
     checkDoor = checkIM();
@@ -295,7 +307,9 @@ void state3_stillness() {
         publishStateTransition(3, 3, checkDoor.doorStatus, checkINS.iAverage);
         saveStateChangeOrAlert(3, 5);
         Log.error("Stillness Alert!!");
-        Particle.publish("Stillness Alert", "stillness alert!!!", PRIVATE);
+        number_of_alerts_published += 1;  // increment the number of alerts published
+        snprintf(alertMessage, sizeof(alertMessage), "{\"numberOfAlertsPublished\": %lu}", number_of_alerts_published);
+        Particle.publish("Stillness Alert", alertMessage, PRIVATE);
         hasStillnessAlertBeenSent = true;
         state3_stillness_timer = millis();                     // reset the stillness timer
         max_stillness_time = &state3_max_long_stillness_time;  // set to compare against the longer stillness threshold for the rest of this session
@@ -306,7 +320,9 @@ void state3_stillness() {
         publishStateTransition(3, 3, checkDoor.doorStatus, checkINS.iAverage);
         saveStateChangeOrAlert(3, 4);
         Log.error("Duration Alert!!");
-        Particle.publish("Duration Alert", "duration alert", PRIVATE);
+        number_of_alerts_published += 1;  // increment the number of alerts published
+        snprintf(alertMessage, sizeof(alertMessage), "{\"numberOfAlertsPublished\": %lu}", number_of_alerts_published);
+        Particle.publish("Duration Alert", alertMessage, PRIVATE);
         hasDurationAlertBeenSent = true;
         state2_duration_timer = millis();
         stateHandler = state3_stillness;